requirements.txt
absl-py==0.9.0
asgiref==3.2.3
astor==0.8.1
astroid==2.3.3
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
celery==4.4.0
certifi==2019.11.28
cffi==1.13.2
click==7.1.1
cryptography==2.8
cycler==0.10.0
Cython==0.29.15
DateTime==4.3
decorator==4.4.2
defusedxml==0.6.0
djangorestframework==3.11.0
django-webpack-loader==0.6.0
dlib==19.19.0
entrypoints==0.3
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
gast==0.2.2
get==2019.4.13
google-pasta==0.2.0
grpcio==1.27.2
gunicorn
h5py==2.10.0
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4
jedi==0.16.0
Jinja2==2.11.1
json5==0.9.2
jsonschema==3.2.0
jupyter-client==6.0.0
jupyter-core==4.6.1
jupyterlab==1.2.6
jupyterlab-server==1.0.7
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==18.0.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
libarchive-c==2.8
llvmlite==0.28.0
locket==0.2.0
lxml==4.3.2
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.6
mkl-random==1.0.1.1
more-itertools==6.0.0
mpmath==1.1.0
msgpack==0.6.1
mtcnn==0.1.0
multipledispatch==0.6.0
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.2
nltk==3.4
nose==1.3.7
notebook==6.0.3
numba==0.43.1
numexpr==2.6.9
numpy==1.18.2
numpydoc==0.8.0
object-detection-0.1==0.1
olefile==0.46
opencv-contrib-python==4.1.2.30
opencv-python==4.2.0.32
openpyxl==2.6.1
opt-einsum==3.2.0
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.6.2
partd==0.3.10
pathlib2==2.3.3
path.py==11.5.0
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pkginfo==1.5.0.1
pluggy==0.9.0
ply==3.11
post==2019.4.13
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.3
psutil==5.6.1
ptyprocess==0.6.0
public==2019.4.13
py==1.8.0
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
Pygments==2.6.1
pylint==2.4.4
PyMySQL==0.9.3
pyparsing==2.4.6
pyrsistent==0.15.7
python-crontab==2.4.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
pyzmq==18.1.1
query-string==2019.4.13
request==2019.4.13
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
sqlparse==0.3.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
wrapt==1.11.1
WTForms==2.2.1
zipp==2.2.0
zope.interface==4.7.1
django-cors-headers
django-celery-results
djangorestframework-simplejwt==4.4.0
psycopg2